build(msvc): clean build emits zero warnings
Third-party dep warnings are now silenced inside dpf.js (it owns the lvgl /
txiki / quickjs / libuv / mbedtls subtrees) and the genuine framework-fork
warnings are fixed at source, so RetroPlug only keeps suppression for the
vendored bits it pulls in itself:
- cmake/sameboy.cmake: -w on the SameBoy core (vendored upstream).
- test: /w on the vendored liblsdj sources; and fix the two real C4312 sentinel
  casts in CommandQueue/EventQueueTests (cast the 32-bit literal via uintptr_t).

Bumps deps/dpf.js (third-party suppression + fork fixes).

+++ packages/native/test/CommandQueueTests.cpp
@@ -72,7 +72,7 @@ TEST_CASE("CommandQueue carries LoadRom payload", "[CommandQueue]") {
     CommandQueue q;
     // We don't actually dereference the pointer in this test — just confirm
     // the variant payload round-trips intact.
-    auto* sentinel = reinterpret_cast<SystemBase*>(0xDEADBEEF);
+    auto* sentinel = reinterpret_cast<SystemBase*>(uintptr_t(0xDEADBEEF));
     REQUIRE(q.tryPush(Command::makeLoadRom(sentinel)));
 
     Command out;
